core: refactor
