--- backend/src/api/mod.rs
@@ -1,3 +1,0 @@
-pub mod bot;
-pub mod microservices;
-pub mod web;
