+++ backend/src/bot/mod.rs
