+++ backend/src/bot/models.rs
