+++ backend/src/bot/photo.rs
