+++ backend/src/bot/redis.rs
