+++ backend/src/web/locks.rs
@@ -8,7 +8,7 @@ use super::{
 };
 use crate::{
     AppError, AppState,
-    api::microservices::redis::{insert_id, remove_id, try_get},
+    microservices::redis::{insert_id, remove_id, try_get},
 };
 use chrono::{Duration as chronoDuration, Utc};
 use once_cell::sync::Lazy;
