+++ backend/src/web/sessions.rs
@@ -9,7 +9,7 @@ use super::{
 };
 use crate::{
     AppError, AppState,
-    api::microservices::redis::{insert_id, remove_id},
+    microservices::redis::{insert_id, remove_id},
 };
 use axum::http::header::HeaderMap;
 use chrono::Utc;
