+++ backend/src/web/swap/database.rs
@@ -1,10 +1,8 @@
 use super::models::{Condition, Emoji, Item, ItemPayload, ItemRow, ItemType, Location};
 use crate::{
     AppError, AppState,
-    api::{
-        microservices::cdc::convert_i8_to_u8,
-        web::models::{RedisAccount, WebsitePath},
-    },
+    microservices::cdc::convert_i8_to_u8,
+    web::models::{RedisAccount, WebsitePath},
 };
 use chrono::{Duration as chronoDuration, Utc};
 use scylla::response::{PagingState, query_result::FirstRowError::RowsEmpty};
