+++ backend/src/web/swap/handlers.rs
@@ -1,5 +1,5 @@
 use super::{models::ItemPayload, redis::try_post_item, utilities::check_item};
-use crate::{AppError, AppState, RedisAction, WebsitePath, api::web::verify::check_token};
+use crate::{AppError, AppState, RedisAction, WebsitePath, web::verify::check_token};
 use axum::{
     Json,
     extract::State,
