+++ backend/src/web/utilities.rs
@@ -1,5 +1,5 @@
 use super::models::{RedisAction, VerifiedTokenResult, WebsitePath};
-use crate::{AppError, AppState, WebsiteRoute, api::microservices::redis::try_get};
+use crate::{AppError, AppState, WebsiteRoute, microservices::redis::try_get};
 use axum::{extract::Request, http::header::HeaderMap};
 use sha2::{Digest, Sha256};
 use std::{net::IpAddr, sync::Arc};
